==> 01_price_data_fetcher_sketches.ipynb <==
import numpy as np
import pandas as pd
import yfinance as yf
import datetime

datetime.date.today()

df = yf.download(["SPY"], start="1900-01-01", end=datetime.date.today(), auto_adjust=False)
df = df["Adj Close"].reset_index()
df["Date"] = df["Date"].dt.date
df.columns = df.columns.str.lower()
df

class PriceFetcher:
    def __init__(self, tickers):
        self.tickers = tickers

    prices = None
    
    def fetch_data(self):
        self.prices = yf.download(self.tickers, start="1900-01-01", end=datetime.date.today(), auto_adjust=False)
        self.prices = self.prices["Adj Close"].reset_index()
        self.prices["Date"] = self.prices["Date"].dt.date
        self.prices.columns = self.prices.columns.str.lower()
        return(None)

pf = PriceFetcher(["SPY", "AGG"])

pf.fetch_data()

pf.prices
