==> 03_market_corrections_example_code.ipynb <==
## Importing Packages

import numpy as np
import pandas as pd

## Reading-In and Wrangling Data

df_spy = pd.read_excel("..\\data\\bufr_bufd_mquslblr.xlsx", "spy")
df_spy

## Calculating Returns and Equity Curves

df_spy["ret_spy"] = df_spy["spy"].pct_change()
df_spy.fillna(0, inplace=True)

df_spy["equity_spy"] = (1 + df_spy["ret_spy"]).cumprod()
df_spy["drawdown_spy"] = (df_spy["equity_spy"] / df_spy["equity_spy"].cummax()) - 1
df_spy

## Determining Drawdown Period Start and End Dates

Notice that this is a bit misleading, because two days of consecutive all time highs will show up as a "drawdown period".  However, when a I do the query for actual corrections, this will be take care of.

start_dates = df_spy.query("drawdown_spy == 0")["date"].iloc[:-1].values
end_dates = df_spy.query("drawdown_spy == 0")["date"].iloc[1:]

df_drawdown_periods = pd.DataFrame({
    "start_date":start_dates,
    "end_date":end_dates,
}).reset_index(drop=True)
df_drawdown_periods

## Calculating the Maximum Drawdowns during Drawdown Period

period_drawdowns = []
bottom_dates = []
for ix in df_drawdown_periods.index:
    start_date = df_drawdown_periods.at[ix, "start_date"]
    end_date = df_drawdown_periods.at[ix, "end_date"]
    drawdown = df_spy.query("@start_date <= date & date <= @end_date")["drawdown_spy"].min()
    df = df_spy.query("@start_date <= date & date <= @end_date")
    bottom_dt = df[df["drawdown_spy"] == drawdown]["date"].iloc[0]
    period_drawdowns.append(drawdown)
    bottom_dates.append(bottom_dt)

df_drawdown_periods["bottom_date"] = bottom_dates
df_drawdown_periods["drawdown_spy"] = period_drawdowns

## Filtering for 5% Corrections

df_corrections = df_drawdown_periods.query("drawdown_spy < -0.05")
df_corrections

df_corrections.to_csv("spy_corrections.csv", index=False)
