==> 04_market_corrections_sketches.ipynb <==
import yfinance as yf
import datetime

from MarketCorrections import MarketCorrections

mc = MarketCorrections(asset="GLD", correction=-0.05)

mc.prices

mc.drawdown_periods

mc.corrections

#mc.corrections["start_date"].iloc[0].strftime("%Y-%m-%d")
